#51 : Abstract away the requesting of notifications permission
Because I need to ask for them in the Home screen as well and I need to keep my code DRY.
