+++ app/screens/Settings/AppSettings/AppSettings.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import { View, Text, Switch } from 'react-native';
-import { Permissions, Notifications } from 'expo';
 import { Content, Right, Left, Body, Icon, List, ListItem } from 'native-base';
 import { connect } from 'react-redux';
 import { auth, database } from 'app/config/firebase';
+import Permissions from 'app/lib/Permissions';
 import ScreenWrapper from 'app/components/common/ScreenWrapper';
 import styling from 'app/config/styling';
 import style from '../style';
@@ -16,7 +16,6 @@ class AppSettings extends Component {
 
 		this.state = this.getInitialState();
 
-		this.askForNotifications = this.askForNotifications.bind(this);
 		this.handleNotificationChange = this.handleNotificationChange.bind(this);
 	}
 
@@ -29,29 +28,9 @@ class AppSettings extends Component {
 		};
 	}
 
-	async askForNotifications() {
-		const {status} = await Permissions.askAsync(Permissions.NOTIFICATIONS);
-		
-		if(status != 'granted') {
-			console.log(status)
-			alert('Please enable notifications, they are essential for the proper functioning of the app.');
-			return;
-		}
-
-		const token = await Notifications.getExpoPushTokenAsync();
-		const Users = database.collection('Users');
-		const {id} = this.props.user;
-
-		Users.doc(id).update({
-			"permissions.notifications": status == 'granted',
-			"permissions.notifications_token": token,
-			updated_at: new Date(),
-		});
-	}
-
 	async handleNotificationChange(activate) {
 		if(activate) {
-			this.askForNotifications();
+			Permissions.requestNotifications(this.props.user.id);
 			return;
 		}
